Add files via upload
This semester-long project takes a closer look at the relationship between natural disasters and housing rates, as well as natural disasters and Insurance rates, in the United States.
This was a group project, where Emma mainly worked on the abstract, part of the introduction, part of the literature review, and the majority of the methods and results specifically on storms. 
Much of the coding and graphs are done using python programming In a separate file not Include here.
The paper and copies of the analyzed datasets are provided to the right.